Reorganised into hook folders with index scripts
