+++ scripts/lib/gpt-api.js
@@ -1,13 +1,13 @@
-import { moduleName, getGamePromptSetting } from './settings.js';
-import { pushHistory } from './history.js';
+import { moduleName, getGamePromptSetting } from '../settings.js';
+import { pushHistory } from '../history.js';
 
 async function callGptApi(query) {
 	const apiKey = game.settings.get(moduleName, 'apiKey');
 	const model = game.settings.get(moduleName, 'modelVersion');
 	const prompt = getGamePromptSetting();
 	const apiUrl = 'https://api.openai.com/v1/chat/completions';
-	const promptMessage = {role: 'user', content: prompt};
-	const queryMessage = {role: 'user', content: query};
+	const promptMessage = { role: 'user', content: prompt };
+	const queryMessage = { role: 'user', content: query };
 	const messages = pushHistory().concat(promptMessage, queryMessage);
 
 	const requestBody = {
