+++ scripts/lib/hooks/index.js
@@ -1,0 +1,3 @@
+export * as chatMessage from './chatMessage/chatMessage.js';
+export * as rollAttack from './rollAttack/rollAttack.js';
+export * as rollDamage from './rollDamage/rollDamage.js';
