+++ scripts/lib/hooks/rollDamage/rollDamage.js
@@ -1,0 +1,4 @@
+import { createPrompt } from "./rollDamage.prompt.js";
+
+// TODO - Main function for rollDamage hook
+export async function main() { }
