+++ scripts/lib/hooks/rollDamage/rollDamage.prompt.js
@@ -1,0 +1,4 @@
+import * as utils from '/modules/gpt-flavor-text/scripts/lib/utils.js';
+
+// TODO - createPrompt for rollDamage hook
+export const createPrompt = () => { }
