+++ scripts/lib/index.js
@@ -1,0 +1,5 @@
+export * as chat from './chat.js';
+export * as gptApi from './gpt-api.js';
+export * as hooks from './hooks/index.js';
+export * as lib from './lib.js';
+export * as utils from './utils.js';
